# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Twiga Travels & Tours API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/me:
    get:
      operationId: adminGetMe
      tags: [admin-auth]
      summary: Get current admin user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/change-password:
    post:
      operationId: adminChangePassword
      tags: [admin-auth]
      summary: Change admin password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPasswordChange"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Dashboard ───────────────────────────────────────────────────────
...
  /admin/destinations:
    get:
      operationId: adminListDestinations
      tags: [admin-destinations]
      summary: List all destinations (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Destinations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDestination"
    post:
      operationId: adminCreateDestination
      tags: [admin-destinations]
      summary: Create a destination
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDestinationInput"
      responses:
        "201":
          description: Destination created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDestination"
  /admin/destinations/{id}:
    patch:
      operationId: adminUpdateDestination
      tags: [admin-destinations]
      summary: Update a destination
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDestinationUpdate"
      responses:
        "200":
          description: Destination updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDestination"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteDestination
      tags: [admin-destinations]
      summary: Delete a destination
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Admin Bookings ────────────────────────────────────────────────────────
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin-bookings]
      summary: List all bookings (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /admin/bookings/{id}/status:
    patch:
      operationId: adminUpdateBookingStatus
      tags: [admin-bookings]
      summary: Update booking status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBookingStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Inquiries ───────────────────────────────────────────────────────
...
  /admin/gallery:
    get:
      operationId: adminListGalleryImages
      tags: [admin-gallery]
      summary: List gallery images
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminGalleryImage"
    post:
      operationId: adminCreateGalleryImage
      tags: [admin-gallery]
      summary: Add a gallery image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGalleryImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGalleryImage"
  /admin/gallery/{id}:
    delete:
      operationId: adminDeleteGalleryImage
      tags: [admin-gallery]
      summary: Delete a gallery image
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Admin Testimonials ────────────────────────────────────────────────────
...
  /admin/blog:
    get:
      operationId: adminListBlogPosts
      tags: [admin-blog]
      summary: List blog posts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBlogPost"
    post:
      operationId: adminCreateBlogPost
      tags: [admin-blog]
      summary: Create a blog post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBlogPost"
  /admin/blog/{id}:
    get:
      operationId: adminGetBlogPost
      tags: [admin-blog]
      summary: Get a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBlogPost"
    patch:
      operationId: adminUpdateBlogPost
      tags: [admin-blog]
      summary: Update a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBlogPost"
    delete:
      operationId: adminDeleteBlogPost
      tags: [admin-blog]
      summary: Delete a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Admin FAQs ────────────────────────────────────────────────────────────
  /admin/faqs:
    get:
      operationId: adminListFaqs
      tags: [admin-faqs]
      summary: List FAQs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminFaq"
    post:
      operationId: adminCreateFaq
      tags: [admin-faqs]
      summary: Create a FAQ
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFaq"
  /admin/faqs/{id}:
    patch:
      operationId: adminUpdateFaq
      tags: [admin-faqs]
      summary: Update a FAQ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFaq"
    delete:
      operationId: adminDeleteFaq
      tags: [admin-faqs]
      summary: Delete a FAQ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Admin Team ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Booking:
      type: object
      required: [id, full_name, email, phone, travel_start_date, num_travelers, status, created_at]
      properties:
        id: { type: string }
        tour_id: { type: ["string", "null"] }
        tour_title: { type: ["string", "null"] }
        full_name: { type: string }
        email: { type: string }
        phone: { type: string }
        travel_start_date: { type: string }
        travel_end_date: { type: ["string", "null"] }
        num_travelers: { type: integer }
        accommodation_level: { type: ["string", "null"] }
        special_requests: { type: ["string", "null"] }
        add_ons:
          type: array
          items: { type: string }
        status: { type: string }
        created_at: { type: string }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Admin Auth Schemas ────────────────────────────────────────────────────
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AdminAuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/AdminUserProfile"
    AdminUserProfile:
      type: object
      required: [id, name, email, role, is_active, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        is_active: { type: boolean }
        created_at: { type: string }
    AdminPasswordChange:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }

    # ── Admin Dashboard Schemas ───────────────────────────────────────────────
...
    AdminDestination:
      type: object
      required: [id, name, slug, country, is_featured, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        image_url: { type: ["string", "null"] }
        country: { type: string }
        region: { type: ["string", "null"] }
        is_featured: { type: boolean }
        created_at: { type: string }
    AdminDestinationInput:
      type: object
      required: [name, slug, country]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        image_url: { type: string }
        country: { type: string }
        region: { type: string }
        is_featured: { type: boolean }
    AdminDestinationUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        image_url: { type: string }
        country: { type: string }
        region: { type: string }
        is_featured: { type: boolean }

    # ── Admin Booking Schemas ─────────────────────────────────────────────────
    AdminBookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }

    # ── Admin Gallery Schemas ─────────────────────────────────────────────────
    AdminGalleryImage:
      type: object
      required: [id, url, category, sort_order, created_at]
      properties:
        id: { type: string }
        url: { type: string }
        caption: { type: ["string", "null"] }
        category: { type: string }
        sort_order: { type: integer }
        created_at: { type: string }
    AdminGalleryImageInput:
      type: object
      required: [url]
      properties:
        url: { type: string, minLength: 1 }
        caption: { type: string }
        category: { type: string }
        sort_order: { type: integer }

    # ── Admin Testimonial Schemas ─────────────────────────────────────────────
...
    AdminBlogPost:
      type: object
      required: [id, title, slug, content, status, tags, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: ["string", "null"] }
        featured_image: { type: ["string", "null"] }
        status: { type: string }
        category: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        author_id: { type: ["string", "null"] }
        created_at: { type: string }
        updated_at: { type: string }
    AdminBlogPostInput:
      type: object
      required: [title, slug, content]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        content: { type: string }
        excerpt: { type: string }
        featured_image: { type: string }
        status: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
    AdminBlogPostUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        featured_image: { type: string }
        status: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }

    # ── Admin FAQ Schemas ─────────────────────────────────────────────────────
    AdminFaq:
      type: object
      required: [id, question, answer, sort_order, is_published, created_at]
      properties:
        id: { type: string }
        question: { type: string }
        answer: { type: string }
        category: { type: ["string", "null"] }
        sort_order: { type: integer }
        is_published: { type: boolean }
        created_at: { type: string }
    AdminFaqInput:
      type: object
      required: [question, answer]
      properties:
        question: { type: string, minLength: 1 }
        answer: { type: string, minLength: 1 }
        category: { type: string }
        sort_order: { type: integer }
        is_published: { type: boolean }
    AdminFaqUpdate:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
        category: { type: string }
        sort_order: { type: integer }
        is_published: { type: boolean }

    # ── Admin Team Schemas ────────────────────────────────────────────────────
...