# lib/api-zod/src/generated/types/adminBlogPostUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Twiga Travels & Tours API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminBlogPostUpdate {
  title?: string;
  slug?: string;
  content?: string;
  excerpt?: string;
  featured_image?: string;
  status?: string;
  category?: string;
  tags?: string[];
}