# lib/api-zod/src/generated/types/adminBookingStatusUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Twiga Travels & Tours API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminBookingStatusUpdate {
  status: string;
}