# lib/api-zod/src/generated/types/adminFaq.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Twiga Travels & Tours API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminFaq {
  id: string;
  question: string;
  answer: string;
  /** @nullable */
  category?: string | null;
  sort_order: number;
  is_published: boolean;
  created_at: string;
}